Add files via upload
this is a csv file separated by 'space', and has 3 features (age, agegrp,CHD) and 100 samples, hence it is a 100x3 matrix. we use this in logistic regression to predict if person has HeartDisease or not. You can see explanation of various statistics, here,  https://www.scipress.io/editpost/nTZIuduECl0j5ol3HnuA